requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
crispy-tailwind==1.0.3
cryptography==45.0.5
distlib==0.3.9
Django==5.2.4
django-allauth==65.10.0
django-browser-reload==1.18.0
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
filelock==3.18.0
greenlet==3.2.3
idna==3.10
mando==0.7.1
mypy==1.14.1
mypy-extensions==1.0.0
mypy-lang==0.5.0
numpy==2.2.1
oauthlib==3.3.1
pandas==2.2.3
pep8==1.7.1
pillow==11.3.0
platformdirs==4.3.8
pycparser==2.22
pyenv-win==3.1.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
radon==6.0.1
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.5.0
virtualenv==20.31.2
virtualenvwrapper-win==1.2.7
